requirements.txt
argcomplete==1.8.2
asn1crypto==0.23.0
attrs==17.3.0
Automat==0.6.0
beautifulsoup4==4.5.3
certifi==2017.11.5
cffi==1.11.2
chardet==2.3.0
constantly==15.1.0
cryptography==2.1.3
cssselect==1.0.1
Cython==0.27.3
docx2txt==0.6
EbookLib==0.15
enum34==1.1.6
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
ipaddress==1.0.18
lxml==4.1.1
olefile==0.44
parsel==1.2.0
Pillow==4.3.0
pkg-resources==0.0.0
pyasn1==0.3.7
pyasn1-modules==0.1.5
pycparser==2.18
PyDispatcher==2.0.5
pyocr==0.4.7
pyOpenSSL==17.3.0
pytesseract==0.1.7
python-pptx==0.6.5
queuelib==1.4.2
requests==2.18.4
Scrapy==1.4.0
service-identity==17.0.0
six==1.10.0
SpeechRecognition==3.6.3
Twisted==17.9.0
urllib3==1.22
w3lib==1.18.0
Wand==0.4.4
xlrd==1.0.0
XlsxWriter==1.0.2
zope.interface==4.4.3
